+++ NEWS.md
@@ -12,11 +12,13 @@ These changes are technically breaking changes, but they should only affect adva
 - Do not rehash large imported files every `make()` ([#878](https://github.com/ropensci/drake/issues/878)).
 - Repair parsing of long tidy eval inputs in the DSL ([#878](https://github.com/ropensci/drake/issues/881)).
 - Clear up cache confusion when a custom cache exists adjacent to the default cache ([#883](https://github.com/ropensci/drake/issues/883)).
+- Accept targets as symbols in `r_drake_build()`.
 
 ## New features
 
 - Add `drake_slice()` to help split data across multiple targets. Related: [#77](https://github.com/ropensci/drake/issues/77), [#685](https://github.com/ropensci/drake/issues/685), [#833](https://github.com/ropensci/drake/issues/833).
 - Introduce a new `drake_cache()` function, which is now recommended instead of `get_cache()` ([#883](https://github.com/ropensci/drake/issues/883)).
+- Introduce a new `r_deps_target()` function.
 
 ## Enhancements